Ported and enabled Os Tests on Windows.
TEST_F(OsTest, Libraries)
The test failed because no environmental variable was provided for
Windows. I used "PATH" as environmental variable on Windows since
there's no default variable where the linker should look into while
linking dynamic libraries/shared libraries. (like LD_LIBRARY_PATH on Unix/Linux)

TEST_F(OsTest, Sleep)
The test hanged because of the last assertion:
ASSERT_ERROR(os::sleep(Milliseconds(-10)));
Apparently, it didn't handle a negative sleep duration. By simply
returning an error value when the duration is negative the problem was
solved.

TEST_F(OsTest, SYMLINK_Size)
First issue was this assertion:
EXPECT_SOME_EQ(size, os::stat::size(file,FollowSymlink::DO_NOT_FOLLOW_SYMLINK));
It fails everytime because <file> is not a symbolic link. When calling
os:stat:size with argument <follow> set to "DO_NOT_FOLLOW_SYMLINK", it firstly
checks whether the path it receives is a symlink in order to get the symlink data
for that path. If it's not a symlink, it returns an error.
If you are supposed to be able to test the size of a non-link file in
the context of not-following-a-link, the program should return the
size of the non-link file instead of returning an error upon determining the
nature of the path as not being a link file path. Right now, the assertion
is not relevant since you'll get an error if you supply it a non-link file
path rather thana symlink.

TEST_F_TEMP_DISABLED_ON_WINDOWS(OsTest, SYMLINK_Realpath)
The issue was that the symlink resolved to itself but it was tested
against the path of the target file. Once the path of the target file
was replaced by the path of the symlink, the test passed.

Review: https://reviews.apache.org/r/63245/